refactor: modularize more game-data files
