+++ js/game-data/talent-renames.js
@@ -1,0 +1,39 @@
+// if a talent was renamed but not removed, store those IDs manually here.
+// will have to monitor heroes talents repo to discover these.
+const TalentRenames = {
+  RehgarMasteryColossalTotem: 'RehgarEarthbindTotemColossalTotem',
+  RehgarMasteryElectricCharge: 'RehgarLightningShieldElectricCharge',
+  RehgarMasteryWolfRun: 'RehgarWolfRun',
+  RehgarMasteryFeralHeart: 'RehgarFeralHeart',
+  RehgarMasteryStormcaller: 'RehgarLightningShieldStormcaller',
+  RehgarMasteryShamanHealingWard: 'RehgarHealingTotem',
+  RehgarGhostWolfBloodAndThunder: 'RehgarBloodAndThunder',
+  RehgarMasteryFarsight: 'RehgarFarsight',
+  RehgarHeroicAbilityAncestralHealing: 'RehgarAncestralHealing',
+  RehgarHeroicAbilityBloodlust: 'RehgarBloodlust',
+  RehgarMasteryEarthShield: 'RehgarLightningShieldEarthShield',
+  RehgarMasteryEarthGraspTotem: 'RehgarEarthbindTotemEarthgraspTotem',
+  RehgarHungeroftheWolf: 'RehgarHungerOfTheWolf',
+  RehgarRisingStorm: 'RehgarLightningShieldRisingStorm',
+  RehgarMasteryFarseersBlessing: 'RehgarAncestralHealingFarseersBlessing',
+  RehgarMasteryGladiatorsWarShout: 'RehgarGladiatorsWarShout',
+  BarbarianMasteryShatteredGroundSeismicSlam: 'BarbarianShatteredGround',
+  BarbarianMasteryHurricaneWhirlwind: 'BarbarianHurricane',
+  BarbarianMasteryWarPaint: 'BarbarianWarPaint',
+  BarbarianMasteryPoisonedSpearAncientSpear: 'BarbarianPoisonedSpear',
+  BarbarianCombatStyleShotofFury: 'BarbarianShotOfFury',
+  BarbarianMasteryMysticalSpearAncientSpear: 'BarbarianMysticalSpear',
+  BarbarianMasteryFuriousBlowSeismicSlam: 'BarbarianGiantSlammer',
+  SonyaTalentNervesOfSteel: 'BarbarianNervesOfSteel',
+  BarbarianMasteryArreatCraterLeap: 'BarbarianArreatCrater',
+  SonyaTalentIgnorePain: 'BarbarianIgnorePain',
+  BrightwingHyperShiftPhaseShift: 'BrightwingPhaseShiftHyperShift',
+  BrightwingPeekabooPhaseShift: 'BrightwingPhaseShiftPeekaboo',
+  FaerieDragonMasteryPhaseShield: 'BrightwingPhaseShiftPhaseShield',
+  RaynorMasteryHelsAngelsRaynorsBanshees: 'RaynorMasteryDuskWingsRaynorsBanshees',
+  JunkratHeroicAbilityRIPTire: 'JunkratRIPTire',
+  JunkratHeroicRocketRide: 'JunkratRocketRide',
+  JunkratSpreadVolley: 'JunkratFragLauncherSpreadVolley'
+};
+
+module.exports = TalentRenames;
